kafka-producer/requirements.txt
kafka-python==2.0.2
pandas==1.5.3
numpy==1.21.6
psycopg2-binary==2.9.6